tests/wpt/tests/service-workers/service-worker/tentative/static-router/static-router-invalid-rules.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>
  Static Router: registration of invalid rules should raise.
</title>
<script src="/common/get-host-info.sub.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/service-workers/service-worker/resources/test-helpers.sub.js">
</script>
<script src="resources/static-router-helpers.sub.js">
</script>
<body>
<script>

const ROUTER_RULE_KEY_INVALID_REQUEST_METHOD =
    'condition-invalid-request-method';
const ROUTER_RULE_KEY_INVALID_OR_CONDITION_DEPTH =
    'condition-invalid-or-condition-depth';
const ROUTER_RULE_KEY_INVALID_ROUTER_SIZE =
    'condition-invalid-router-size';

promise_test(async t => {
  const worker = await registerAndActivate(t, ROUTER_RULE_KEY_INVALID_REQUEST_METHOD);
  t.add_cleanup(() => {reset_info_in_worker(worker)});
  const {errors} = await get_info_from_worker(worker);
  assert_equals(errors.length, 1);
}, 'addRoutes should raise for invalid request method.');

promise_test(async t => {
  const worker = await registerAndActivate(t, ROUTER_RULE_KEY_INVALID_OR_CONDITION_DEPTH);
  t.add_cleanup(() => {reset_info_in_worker(worker)});
  const {errors} = await get_info_from_worker(worker);
  assert_equals(errors.length, 1);
}, 'addRoutes should raise if or condition exceeds the depth limit');

promise_test(async t => {
  const worker = await registerAndActivate(t, ROUTER_RULE_KEY_INVALID_ROUTER_SIZE);
  t.add_cleanup(() => {reset_info_in_worker(worker)});
  const {errors} = await get_info_from_worker(worker);
  assert_equals(errors.length, 1);
}, 'addRoutes should raise if the number of router rules exceeds the length limit');

</script>
</body>
